+++ docs/news.rst
@@ -3,6 +3,7 @@ Release Notes
 
 **UNRELEASED**
 
+- Updated vendored ``packaging`` to 23.2
 - Fixed ABI tag generation for CPython 3.13a1 on Windows (PR by Sam Gross)
 
 **0.41.2 (2023-08-22)**